+++ platform/win32/functions/threads/mtx_trylock.c
@@ -6,7 +6,7 @@ int mtx_trylock(mtx_t *mtx)
 {
     DWORD myId = GetCurrentThreadId();
 
-    if(mtx->_ThreadId == myId) {
+    if(mtx->_ThreadId == (long) myId) {
         mtx->_NestCount++;
         return thrd_success;
     }
